small bugfix

+++ D2/src/D2.cpp
@@ -658,7 +658,7 @@ vector<D2SpecLine> getLocalMinima(const vector<D2SpecLine>& spec, int maxCount) 
 
 double getError(const D2SpecLine& minimum, const vector<D2SpecLine>& spec) {
 	//cout << "Conf. int. for minimum: " << minimum.frequency << " " << minimum.tyv << " " << minimum.tav << endl;
-	size_t index = 0;
+	int index = 0;
 	for (auto s : spec) {
 		if (s.frequency == minimum.frequency) {
 			break;
@@ -668,7 +668,7 @@ double getError(const D2SpecLine& minimum, const vector<D2SpecLine>& spec) {
 	double sum = 0;
 	double sum2 = 0;
 	if (index < spec.size()) {
-		for (size_t i = index - 1; i >= 0; i--) {
+		for (int i = index - 1; i >= 0; i--) {
 			double probFact = exp(minimum.tav * (1 - minimum.tav * spec[i].tyv / minimum.tyv / spec[i].tav));
 			//cout << "Lower: " << spec[i].frequency << " " << spec[i].tyv << " " << spec[i].tav << " " << probFact << endl;
 			if (probFact < 0.001) {
@@ -677,7 +677,7 @@ double getError(const D2SpecLine& minimum, const vector<D2SpecLine>& spec) {
 			sum += sqrt(-2*log(probFact)) * abs(spec[i].frequency - minimum.frequency);
 			sum2 += -2*log(probFact);
 		}
-		for (size_t i = index + 1; i < spec.size(); i++) {
+		for (int i = index + 1; i < spec.size(); i++) {
 			double probFact = exp(minimum.tav * (1 - minimum.tav * spec[i].tyv / minimum.tyv / spec[i].tav));
 			//cout << "Upper: " << spec[i].frequency << " " << spec[i].tyv << " " << spec[i].tav << " " << probFact << endl;
 			if (probFact < 0.001) {
